fix: plastic LoRA dtype mismatch + compact_memory cache graph leak
Two fixes:
1. plastic_A/plastic_B cast to synch dtype before matmul — bfloat16
   model with float32 LoRA params caused RuntimeError on GPU.

2. compact_memory cache rebuild each step — forward writes into cache
   tensors which become part of computation graph. After backward
   frees them, copy_() into those tensors failed on step 2+. Now
   rebuilds with detach().clone() each iteration.

Key finding: plasticity has a CRITICAL PERIOD. Step 2000 model
accepts compact_memory (4/4 recall). Step 3000+ rejects it — same
settings, same lr, same steps, weights won't budge. Training
hardens the optimization landscape. Needs --plasticity-every to
maintain malleability throughout training.

+++ nanochat/gpt.py
@@ -740,7 +740,7 @@ class CTMBlock(nn.Module):
                 # Plastic LoRA for multi-tick path
                 _gate = torch.sigmoid(self.plastic_gate)
                 if _gate > 1e-4:
-                    tick_out = tick_out + _gate * (synch_k @ self.plastic_A @ self.plastic_B).reshape(B, T, D)
+                    tick_out = tick_out + _gate * (synch_k @ self.plastic_A.to(synch_k.dtype) @ self.plastic_B.to(synch_k.dtype)).reshape(B, T, D)
                 grad_ticks = getattr(self, 'multi_tick_grad', 4)
                 if k < K - grad_ticks:
                     tick_out = tick_out.detach()
@@ -768,7 +768,7 @@ class CTMBlock(nn.Module):
         # Plastic LoRA: additive pathway for compact_memory-learned knowledge
         gate = torch.sigmoid(self.plastic_gate)
         if gate > 1e-4:  # skip when gate is ~0 (untrained)
-            plastic_out = (synch @ self.plastic_A @ self.plastic_B).reshape(B, T, D)
+            plastic_out = (synch @ self.plastic_A.to(synch.dtype) @ self.plastic_B.to(synch.dtype)).reshape(B, T, D)
             out = out + gate * plastic_out
 
         # ROADMAP (phase 5 - metacognitive tokens):
